requirement.txt
flask==3.1.1
flask-cors==6.0.1
chromadb==1.0.12
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.66
langchain-google-genai==2.1.5
langchain-text-splitters==0.3.8
python-dotenv==1.1.0
protobuf==3.20.3
